+++ src/core/token.js
@@ -1,1 +1,1 @@
-export const token = "JSESSIONID=3daf5e1b-5d40-4338-99f4-6ff45ae29088";
+export const token = "JSESSIONID=80826b04-17d2-4ddd-a576-ffb3b3a3b6c5";
